requirements.txt
asgiref==3.2.7
astroid==2.3.3
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==2.1.9
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.9
inflection==0.3.1
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.17.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.6
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
wrapt==1.11.2
